--- src/app/@modal/(.)search/default.tsx
@@ -1,6 +1,0 @@
-// NOTE: /の時にmodalスロットのパラレルルートが404にならないように設定
-const defaultPage = () => {
-  return null;
-}
-
-export default defaultPage;
