--- src/app/@modal/(.)search/page.tsx
@@ -1,12 +1,0 @@
-// 検索モーダル
-import  Modal  from "@/components/Modal";
-
-const searchPage = () => {
-  return (
-    <Modal>
-      <div>検索条件</div>
-    </Modal>
-  );
-}
-
-export default searchPage;
