+++ src/app/search/@modal/default.tsx
@@ -1,0 +1,7 @@
+//  Parallel Routes
+// NOTE: modalスロットのパラレルルートが404にならないように設定
+const defaultPage = () => {
+  return null;
+}
+
+export default defaultPage;
