+++ src/app/search/layout.tsx
@@ -1,0 +1,10 @@
+const SearchLayout = ({children,modal,}: {children: React.ReactNode;modal: React.ReactNode;}) => {
+  return (
+    <div>
+      {children}
+      {modal}
+    </div>
+  );
+}
+
+export default SearchLayout;
