+++ src/app/search/page.tsx
@@ -1,0 +1,10 @@
+const Search = () => {
+  return(
+    <>
+      {/* 何も条件が指定されていないときの検索結果が表示されるように */}
+      <div>検索結果</div>
+    </>
+  )
+};
+
+export default Search;
